+++ mayan/apps/metadata/tests/mixins/document_metadata_mixins.py
@@ -4,7 +4,7 @@ from mayan.apps.documents.tests.mixins.document_mixins import (
     DocumentTestMixin
 )
 
-from ...models import DocumentMetadata
+from ...models.metadata_instance_models import DocumentMetadata
 
 from ..literals import TEST_METADATA_VALUE, TEST_METADATA_VALUE_EDITED