+++ mayan/apps/metadata/wizard_steps.py
@@ -11,7 +11,7 @@ from mayan.apps.sources.wizard_steps import (
 )
 
 from .forms import DocumentMetadataFormSet
-from .models import MetadataType
+from .models.metadata_type_models import MetadataType
 from .permissions import permission_document_metadata_add
 from .utils import decode_metadata_from_query_string, save_metadata_list